+++ Changelog.md
@@ -3,10 +3,12 @@ AppleALC Changelog
 
 #### v1.0.1
 - boot-args presence is now optional
+- platforms must conform to layout-id from now on
 - Added AD1988B codec support (by Rodion)
 - Added ALC888 codec support (by Vandroiy)
 - Added ALC269VC (Lenovo Z580) codec support (by John)
 - Added ALC889 codec support (by Ukr55)
+- Added ALC269 (Acer Aspire) codec support (by Andrey1970)
 - Added PinConfs plugin injecting pin configurations
 
 #### v1.0.0